requirements.txt
Django==2.2.3
django-redis==4.10.0
djangorestframework==3.9.4
djangorestframework-camel-case==1.0.3
entrypoints==0.3
flake8==3.7.7
flake8-broken-line==0.1.0
flake8-isort==2.7.0
flake8-quotes==2.0.1
isort==4.3.21
mccabe==0.6.1
psycopg2-binary==2.8.3
pycodestyle==2.5.0
pyflakes==2.1.1
python-dotenv==0.10.3
pytz==2019.1
redis==3.2.1
six==1.12.0
sqlparse==0.3.0
testfixtures==6.10.0
